springer_results.yml
This is an example code used to pull search results from a publishing journal's website, example here is Springer Nature. The csv file printed will contain title, author, doi, journal, year, and link for all of the articles found with the specified key words. The used needs to first sign up for an API key through Springer Nature in order to obtain results.